Nests models to reduce file sprawl
