+++ StripeCryptoOnramp/StripeCryptoOnramp/Source/Internal/API Bindings/Additional KYC/AdditionalKYCFulfillmentQuestionnaire.swift
@@ -10,7 +10,23 @@ import Foundation
 /// Questionnaire answers submitted for an additional KYC requirement.
 struct AdditionalKYCFulfillmentQuestionnaire: Codable, Equatable {
 
-    /// The customer's answers to the questionnaire.
-    let answers: [AdditionalKYCQuestionnaireAnswer]
+    /// An answer submitted for one additional KYC question.
+    struct Answer: Codable, Equatable {
 
+        /// The API identifier of the question being answered.
+        let questionId: String
+
+        /// The customer's answer.
+        let value: String
+
+        // MARK: - Codable
+
+        private enum CodingKeys: String, CodingKey {
+            case questionId = "question_id"
+            case value
+        }
+    }
+
+    /// The customer's answers to the questionnaire.
+    let answers: [Answer]
 }
