+++ StripeCryptoOnramp/StripeCryptoOnramp/Source/Internal/API Bindings/Additional KYC/FulfillAdditionalKYCRequirementResponse.swift
@@ -10,6 +10,31 @@ import Foundation
 /// Decodable model returned by the `/v1/crypto/internal/fulfill_additional_kyc_requirement` endpoint.
 struct FulfillAdditionalKYCRequirementResponse: Decodable, Equatable {
 
+    /// A document included in a response after fulfilling an additional KYC requirement.
+    struct Document: Decodable, Equatable {
+
+        /// The general kind of document submitted.
+        let documentType: String
+
+        /// The specific document subtype selected by the customer.
+        let documentSubtype: String?
+
+        /// The identifiers of the uploaded document files.
+        let fileIds: [String]
+
+        /// The document's verification status.
+        let status: String
+
+        // MARK: - Decodable
+
+        private enum CodingKeys: String, CodingKey {
+            case documentType = "document_type"
+            case documentSubtype = "document_subtype"
+            case fileIds = "file_ids"
+            case status
+        }
+    }
+
     /// The unique identifier for the submission.
     let id: String
 
@@ -23,7 +48,7 @@ struct FulfillAdditionalKYCRequirementResponse: Decodable, Equatable {
     let submissionType: AdditionalKYCSubmissionType
 
     /// The documents included in the submission.
-    let documents: [FulfillAdditionalKYCRequirementResponseDocument]?
+    let documents: [Document]?
 
     /// The questionnaire answers included in the submission.
     let questionnaire: AdditionalKYCFulfillmentQuestionnaire?
