+++ StripeCryptoOnramp/StripeCryptoOnrampTests/Unit/STPAPIClient+CryptoOnrampTests.swift
@@ -298,7 +298,7 @@ final class STPAPIClientCryptoOnrampTests: APIStubbedTestCase {
             liquidityProvider: "swapped",
             submissionType: .document,
             documents: [
-                FulfillAdditionalKYCRequirementRequestDocument(
+                FulfillAdditionalKYCRequirementRequest.Document(
                     documentType: "source_of_funds",
                     documentSubtype: "payslip",
                     fileIds: ["file_123"]
@@ -306,7 +306,7 @@ final class STPAPIClientCryptoOnrampTests: APIStubbedTestCase {
             ],
             questionnaire: AdditionalKYCFulfillmentQuestionnaire(
                 answers: [
-                    AdditionalKYCQuestionnaireAnswer(
+                    AdditionalKYCFulfillmentQuestionnaire.Answer(
                         questionId: "purchase_purpose",
                         value: "Long-term investment"
                     ),
